# src/swagger.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: No description
  version: 0.1.9

servers:
  - url: http://localhost:8080
    description: Main (production) server

components:
  schemas:
    ExchangeRate:
      properties:
        currency_from:
          type: string
        currency_to:
          type: string
        value:
          type: integer
      required:
        - currency_from
        - currency_to
        - value

paths:
  /convert:
    get:
      tags:
        - Convert
      summary: Gets a user by ID.
      description: convert currency
      operationId: get_convert_currency
      parameters:
        - in: query
          name: from
          schema:
            type: string
          required: true
        - in: query
          name: to
          schema:
            type: string
          required: true
        - in: query
          name: amount
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Successful operation

  /database:
    post:
      tags:
        - Exchange Rate
      summary: Set new exchange rate
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: merge
          schema:
            type: integer
          required: true
        - in: body
          name: data
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ExchangeRate'
#      parameters:
#        - in: query
#          name: merge
#          schema:
#            type: integer
#          required: true
#      requestBody:
#        description: List of exchange rates
#        required: true
#        content:
#          text/plain:
#            schema:
#              type: string
#          application/json:
#            schema:
#              type: array
#              items:
#                $ref: '#/components/schemas/ExchangeRate'
      responses:
        '201': # status code
          description: A JSON array of new exchange rate